Bug #14298: Fix mode validation to be less ridiculous.
The code sort of assumed you'd asked for a virtual size in the config file,
either explicitly or through asking for a mode.  Autoconfiguration will
default the virtual to zero though, so the constraint will fail, and then
a mode we can't do will get through xf86ValidateModes().

Just move the check to S3VValidMode where it belongs.