Respect rotation upon ShadowFB initialization
fixes ShadowFB initialization while respecting rotation

The shadowFB is initialized using the screen's VirtualX for width and
VirtualY for height and regardless of the rotation.

When using the
  Option "Rotate" "CW"
such initialization of the ShadowFB results in such a setup

331
331
22

although the correct setup would be

33
33
33

the 1s and 3s represent the usable workspace.
the 2s and 3s represent the monitor.

Signed-off-by: Paulo Cesar Pereira de Andrade <pcpa@mandriva.com.br>